Move PeerProvider
