+++ my-react-app/src/connection-manager/constants.js
@@ -1,0 +1,3 @@
+export const messageTypes = {
+    CHANGE_CLIMB: 'change_climb'
+}
